For subcommands using IPC, put the client in cmd/
This splits the client and server sides of these APIs, but they're APIs
we're comfortable with other clients accessing as well, so not something
we change frequently.  The `comp/core/ipc/ipcclient` component makes it
easy enough to fetch the data that this is not a great hardship or
introducing a great deal of logic into the command.