Refactor command-running to run oneshots after startup
This refactors the command-running utilities to take the fx.Options as
variadic arguments, so that fxapps.OneShot can add its own fx.Option.

It uses a bit of reflection to convince Fx to gather the arguments to a
function without actually calling that function immediately, allowing
the oneShot function to run after `app.Start` has finished, when we're
sure that all components have started.  This works around the startup
ordering bug in https://github.com/uber-go/fx/issues/918.